requirements.txt
streamlit
pandas
serpapi
python-dotenv
inquirer